# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse": {
				"properties": {
					"status": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"status",
					"statusCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AssessmentQuestion": {
  "type": "object",
  "required": [
    "questionNumber",
    "subject",
    "class",
    "topic",
    "question",
    "options",
    "answer",
    "difficulty",
    "category"
  ],
  "properties": {
    "questionNumber": { "type": "string", "example": "G001" },
    "subject": { "type": "string", "example": "English" },
    "class": { "type": "string", "example": "JSS2" },
    "topic": { "type": "string", "example": "grammar" },
    "question": { "type": "string" },
    "options": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "key": { "type": "string", "example": "a" },
          "text": { "type": "string" }
        }
      }
    },
    "answer": { "type": "string", "example": "b" },
    "difficulty": {
      "type": "string",
      "enum": ["easy", "medium", "hard"]
    },
    "category": {
      "type": "string",
      "enum": ["grammar", "comprehension", "vocabulary", "oral", "writing"]
    },
    "explanation": { "type": "string" }
  }
},
"AssessmentResponse": {
  "type": "object",
  "properties": {
    "total": { "type": "number", "example": 20 },
    "questions": {
      "type": "array",
      "items": {
        "$ref": "#/components/schemas/AssessmentQuestion"
      }
    }
  }
}
		},
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "learnexo",
		"version": "1.0.0",
		"description": "this is an edtech application for personalized learning using Artificial Intelligence",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "serg lab"
		}
	},
	"paths": {
		"/api/v1/questionnaire/bulk-upload": {
			"post": {
				"operationId": "BulkUploadQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Questionnaire"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/questionnaire/{id}": {
			"put": {
				"operationId": "UpdateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Questionnaire"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Questionnaire"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/assessment/questions/bulk-upload": {
  "post": {
    "tags": ["Assessment"],
    "summary": "Bulk upload assessment questions",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssessmentQuestion"
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Questions uploaded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      }
    }
  }
},
"/api/v1/assessment/{subject}/{class}": {
  "get": {
    "tags": ["Assessment"],
    "summary": "Get assessment questions",
    "parameters": [
      {
        "in": "path",
        "name": "subject",
        "required": true,
        "schema": { "type": "string" }
      },
      {
        "in": "path",
        "name": "class",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "responses": {
      "200": {
        "description": "Assessment fetched",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AssessmentResponse"
            }
          }
        }
      }
    }
  }
},
"/api/v1/assessment/{subject}/{class}/topic/{topic}": {
  "get": {
    "tags": ["Assessment"],
    "summary": "Get questions by topic",
    "parameters": [
      {
        "in": "path",
        "name": "subject",
        "required": true,
        "schema": { "type": "string" }
      },
      {
        "in": "path",
        "name": "class",
        "required": true,
        "schema": { "type": "string" }
      },
      {
        "in": "path",
        "name": "topic",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "responses": {
      "200": {
        "description": "Topic questions fetched",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AssessmentResponse"
            }
          }
        }
      }
    }
  }
},"/api/v1/assessment/questions": {
...
"/api/v1/assessment/questions/{questionNumber}": {
  "patch": {
    "tags": ["Assessment"],
    "summary": "Update a question",
    "parameters": [
      {
        "in": "path",
        "name": "questionNumber",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "$ref": "#/components/schemas/AssessmentQuestion"
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Updated"
      }
    }
  },
  "delete": {
    "tags": ["Assessment"],
    "summary": "Delete a question",
    "parameters": [
      {
        "in": "path",
        "name": "questionNumber",
        "required": true,
        "schema": { "type": "string" }
      }
    ],
    "responses": {
      "200": {
        "description": "Deleted"
      }
    }
  }
}
	},
	"servers": [
		{
			"url": "https://dev-api.learnexo.ng",
			"description": "API Gateway development"
		},
		{
			"url": "http://localhost:7070",
			"description": "API Gateway local"
		}
	],
...
}